custom-edit/requirements.txt
accelerate
torchvision
transformers>=4.25.1
ftfy
tensorboard
Jinja2
wandb
xformers

# p2p_utils.py
ipython
ipywidgets